refactor : 상수를 한 파일로 통일
