+++ src/constants/cardConstants.ts
@@ -1,0 +1,24 @@
+// TODO : constants 합치기
+export const MIN_EXPIRATION_YEAR = 25;
+export const MAX_EXPIRATION_YEAR = 99;
+export const MIN_EXPIRATION_MONTH = 1;
+export const MAX_EXPIRATION_MONTH = 12;
+
+export const cardBrands = [
+  'BC카드',
+  '신한카드',
+  '카카오뱅크',
+  '현대카드',
+  '우리카드',
+  '롯데카드',
+  '하나카드',
+  '국민카드',
+];
+
+const cardNetworkLogos = {
+  visa: './visa.png',
+  mastercard: './mastercard.png',
+  default: './default.png',
+};
+
+export default cardNetworkLogos;
