--- src/constants/cardNetworkLogos.ts
@@ -1,7 +1,0 @@
-const cardNetworkLogos = {
-  visa: './visa.png',
-  mastercard: './mastercard.png',
-  default: './default.png',
-};
-
-export default cardNetworkLogos;
