--- src/constants/expiration.ts
@@ -1,4 +1,0 @@
-export const MIN_EXPIRATION_YEAR = 25;
-export const MAX_EXPIRATION_YEAR = 99;
-export const MIN_EXPIRATION_MONTH = 1;
-export const MAX_EXPIRATION_MONTH = 12;
